Added missing info to the GX output files
The actual name/path EBTEL table used to produce the maps wwas missing fro the GX Simulator output files. Thus, to reproduce te results, one had to guess what EBTEL table, if any, was used.
This info is now added to the gxl, gxc, and map files, so the original table is avaialble when these data products are restored. Particularly, the GX GUI restore such EBTEL path when a gxc file is restored.