Added functionality to send a restored model to a GX instance
gx_miodel2gui.pro may be used to send a restored model object to an existing instance of GX Simulator, or to launch GX Simulator if it is not already open.
The box.id has been adopted as the default name of the model objects created by gx_importmodel.pro function. Use Model->GetName() to retrieve this property
The box.execute script has been addopted as the default identifier of the model objects created by gx_importmodel.pro function. Use Model->GetID() to retrive this property