Updated AIA.pro and related routines
Updated AIA.pro and related routines, to replace the static AIA response function with a time dependent response function that matches the time of the GX model. IN the process, the AIA.pro GUI interface have been also upgraded to make use of the global parameters nparms and rparms, in adition to the legacy parms table.