added update keyword to gx_nlfff_libpath
One may use update to force recompilation of the NLFFF library on Linux/WinOS even if it already exists in the gx_binaries folder. Such action may be required after an upgrade of the package. The update may be forced from the command line by issuing this call
IDL> gx_nlfff_libpath(/update)