added the option of locking the fluxtube ownership over the volume voxels
If more than one fluxtube is created, they compete for volume ownership based on an automatic evaluation of the fluxtubes' pressures, which are proportional to the product of their magnetic field strength and temperature. However, the latest update allows expert users to lock the volume ownership of a specific fluxtube, ensuring that adjacent fluxtubes with higher pressures cannot take over the locked volume.