redefined the Q uncertainty range in terms of the local curvature of the metrics
In previous versions, the Q uncertainty range was calculated as an accuracy range spanning the space between the closest Q neighbors of the solution on the search grid, which was based on the golden ratio. The new definition, however, is derived from the local curvature of the metrics versus Q dependence. This is achieved by approximating the dependence with a quadratic function